Optimize squared skew and added power transform
This is a large overhaul of the function, changing the following main things
(1)  Changed argument 'scale' to 'scaled' to avoid confusion between the argument name and the base function 'scale'.
(2) Added a second option for transformation technique. Now, the user has the ability to choose between a log transform and a power transform. Either will be optimized in the same manner.
(3) Added argument 'retain_domain'. Though not recommended, the user has the option of retaining the domain of the original variable after transformation.
(4) Changed the argument 'kurt_folder' to 'skew_folder' since now the goal is to minimize the skew^2 value.
(5) Removed the automatic conversion of the input data to matrix form. The input data must be in matrix form to start with.
(6) Removed the requirement that input data must be positive. Now, all data is automatically shifted and scaled to the domain [0,1] before transformation.
(7) Added default variable names if none are specified.
(8) Removed logarithmic scaling of transformation variables.
(9) Changed name of var_obs_trans to var_trans.