Fixing minor bugs
(1) In documentation example, corrected the last line so that a histogram is printed for each variable
(2)  Corrected assignment of variable name.
(3) Included domain change in transform function
(4) Corrected optimization of trans_val for the 'log' type.
(5) Removed creation of the skew plot for variables that are not transformed.
(6) Corrected raw histograms